Moved webhook logic to `utils`
