+++ packages/new-server/src/controllers/index.controller.ts
@@ -2,6 +2,7 @@ import { IStandardResponse, SendApiResponse } from '@utils/api.response';
 import { Controller, Get, Route, Tags, Body, Post, Header } from 'tsoa';
 import startAITools from '@aitools/main';
 import crypto from 'crypto';
+import { validateWebhook } from '@utils/validateWebhook';
 
 @Tags('Index')
 @Route('')
@@ -16,34 +17,9 @@ export class IndexController extends Controller {
     @Body() payload: any,
   ): Promise<IStandardResponse<string>> {
     try {
-      const elements = livepeerSignature.split(',');
-      const signatureParts = elements.reduce((acc, element) => {
-        const [key, value] = element.split('=');
-        acc[key] = value;
-        return acc;
-      }, {});
+      const webhookAuth = validateWebhook(livepeerSignature, payload);
 
-      const timestamp = signatureParts['t'];
-      const signature = signatureParts['v1'];
-      const signedPayload = JSON.stringify(payload);
-
-      const secret = process.env.LIVEPEER_WEBHOOK_SECRET;
-      const expectedSignature = crypto
-        .createHmac('sha256', secret)
-        .update(signedPayload)
-        .digest('hex');
-
-      const isSignatureValid = crypto.timingSafeEqual(
-        Buffer.from(signature),
-        Buffer.from(expectedSignature),
-      );
-
-      const tolerance = 8 * 60 * 1000; // 8 minutes in milliseconds
-      const currentTime = Date.now(); // Current time in milliseconds
-      const isTimestampValid =
-        Math.abs(currentTime - parseInt(timestamp, 10)) < tolerance;
-
-      if (!isSignatureValid || !isTimestampValid) {
+      if (!webhookAuth) {
         console.log('Invalid signature or timestamp');
         return SendApiResponse('Invalid signature or timestamp', null, '401');
       }
